requirements_old.txt
matplotlib==3.4.1
openpyxl==3.1.2
numpy==1.20.1
robin_stocks==3.3.0
pandas==1.3.3
pyotp==2.6.0
requests==2.26.0
pyEX==0.5.0
python-dotenv==0.19.0
sendgrid
certifi
postmarker==1.0